+++ Assets/ClickerRoot/Scripts/Utils/EventBus/Signals/ScoreChangeSignal.cs
@@ -1,7 +1,7 @@
 ﻿namespace ClickerRoot.Scripts.Utils.Signals
 {
-    public class ScoreChangeSignal : AbstractSignal<int>
+    public class ScoreChangeSignal : AbstractSignal<ulong>
     {
-        public ScoreChangeSignal(int value) : base(value) { }
+        public ScoreChangeSignal(ulong value) : base(value) { }
     }
 }
